refactor: remove all v2 references from codebase
- Rename lua/spaghetti-comb-v2/ to lua/spaghetti-comb/
- Rename plugin/spaghetti-comb-v2.lua to plugin/spaghetti-comb.lua
- Update all require statements from spaghetti-comb-v2 to spaghetti-comb
- Update documentation files (README, SPEC, etc.) to remove v2 branding
- Rename v2-specific ticket files (009, 010, 011)
- Update global variable from vim.g.loaded_spaghetti_comb_v2 to vim.g.loaded_spaghetti_comb
- Update all user-facing strings and messages
- Keep semantic version numbers (v2.0.0, v2.1, etc.) for future releases
