+++ lua/spaghetti-comb/history/manager.lua
@@ -1,6 +1,6 @@
 -- Core history tracking logic
-local types = require("spaghetti-comb-v2.types")
-local project_utils = require("spaghetti-comb-v2.utils.project")
+local types = require("spaghetti-comb.types")
+local project_utils = require("spaghetti-comb.utils.project")
 
 local M = {}
 
@@ -686,7 +686,7 @@ function M.save_current_project_history()
     if not state.initialized then return false, "History manager not initialized" end
     if not state.current_project then return false, "No current project" end
 
-    local storage = require("spaghetti-comb-v2.history.storage")
+    local storage = require("spaghetti-comb.history.storage")
     local trail = state.trails[state.current_project]
 
     if not trail or #trail.entries == 0 then
@@ -700,7 +700,7 @@ end
 function M.save_all_histories()
     if not state.initialized then return false, "History manager not initialized" end
 
-    local storage = require("spaghetti-comb-v2.history.storage")
+    local storage = require("spaghetti-comb.history.storage")
     local saved_count = 0
     local errors = {}
 
@@ -727,7 +727,7 @@ function M.load_project_history(project_root)
     if not state.initialized then return false, "History manager not initialized" end
     if not project_root then return false, "Invalid project root" end
 
-    local storage = require("spaghetti-comb-v2.history.storage")
+    local storage = require("spaghetti-comb.history.storage")
     local trail, err = storage.load_history(project_root)
 
     if not trail then
@@ -772,7 +772,7 @@ function M.try_load_on_project_switch(project_root)
         return false, "History already loaded"
     end
 
-    local storage = require("spaghetti-comb-v2.history.storage")
+    local storage = require("spaghetti-comb.history.storage")
     local trail, err = storage.load_history(project_root)
 
     if trail then
