+++ lua/spaghetti-comb/navigation/events.lua
@@ -2,9 +2,9 @@
 local M = {}
 
 -- Dependencies
-local history_manager = require("spaghetti-comb-v2.history.manager")
-local project_utils = require("spaghetti-comb-v2.utils.project")
-local debug_utils = require("spaghetti-comb-v2.utils.debug")
+local history_manager = require("spaghetti-comb.history.manager")
+local project_utils = require("spaghetti-comb.utils.project")
+local debug_utils = require("spaghetti-comb.utils.debug")
 
 -- Module state
 local state = {
