+++ lua/spaghetti-comb/navigation/jumplist.lua
@@ -2,8 +2,8 @@
 local M = {}
 
 -- Dependencies
-local history_manager = require("spaghetti-comb-v2.history.manager")
-local debug_utils = require("spaghetti-comb-v2.utils.debug")
+local history_manager = require("spaghetti-comb.history.manager")
+local debug_utils = require("spaghetti-comb.utils.debug")
 
 -- Module state
 local state = {
