+++ lua/spaghetti-comb/tests/history_spec.lua
@@ -1,6 +1,6 @@
 -- History manager tests
 local MiniTest = require("mini.test")
-local history_manager = require("spaghetti-comb-v2.history.manager")
+local history_manager = require("spaghetti-comb.history.manager")
 
 local T = MiniTest.new_set()
