+++ lua/spaghetti-comb/tests/init.lua
@@ -1,4 +1,4 @@
--- Test runner for spaghetti-comb-v2
+-- Test runner for spaghetti-comb
 local M = {}
 
 -- Run all tests
@@ -7,12 +7,12 @@ function M.run_all()
 
     -- Collect all test sets
     local test_sets = {
-        require("spaghetti-comb-v2.tests.history_spec"),
-        require("spaghetti-comb-v2.tests.bookmarks_spec"),
-        require("spaghetti-comb-v2.tests.ui_spec"),
-        require("spaghetti-comb-v2.tests.navigation_spec"),
-        require("spaghetti-comb-v2.tests.integration_spec"),
-        require("spaghetti-comb-v2.tests.lsp_spec"),
+        require("spaghetti-comb.tests.history_spec"),
+        require("spaghetti-comb.tests.bookmarks_spec"),
+        require("spaghetti-comb.tests.ui_spec"),
+        require("spaghetti-comb.tests.navigation_spec"),
+        require("spaghetti-comb.tests.integration_spec"),
+        require("spaghetti-comb.tests.lsp_spec"),
     }
 
     -- Run tests
@@ -24,37 +24,37 @@ end
 -- Run specific test category
 function M.run_history()
     local MiniTest = require("mini.test")
-    local test_set = require("spaghetti-comb-v2.tests.history_spec")
+    local test_set = require("spaghetti-comb.tests.history_spec")
     MiniTest.run(test_set)
 end
 
 function M.run_ui()
     local MiniTest = require("mini.test")
-    local test_set = require("spaghetti-comb-v2.tests.ui_spec")
+    local test_set = require("spaghetti-comb.tests.ui_spec")
     MiniTest.run(test_set)
 end
 
 function M.run_navigation()
     local MiniTest = require("mini.test")
-    local test_set = require("spaghetti-comb-v2.tests.navigation_spec")
+    local test_set = require("spaghetti-comb.tests.navigation_spec")
     MiniTest.run(test_set)
 end
 
 function M.run_integration()
     local MiniTest = require("mini.test")
-    local test_set = require("spaghetti-comb-v2.tests.integration_spec")
+    local test_set = require("spaghetti-comb.tests.integration_spec")
     MiniTest.run(test_set)
 end
 
 function M.run_lsp()
     local MiniTest = require("mini.test")
-    local test_set = require("spaghetti-comb-v2.tests.lsp_spec")
+    local test_set = require("spaghetti-comb.tests.lsp_spec")
     MiniTest.run(test_set)
 end
 
 function M.run_bookmarks()
     local MiniTest = require("mini.test")
-    local test_set = require("spaghetti-comb-v2.tests.bookmarks_spec")
+    local test_set = require("spaghetti-comb.tests.bookmarks_spec")
     MiniTest.run(test_set)
 end
