+++ plugin/spaghetti-comb.lua
@@ -1,30 +1,30 @@
--- Spaghetti Comb v2 - Neovim Plugin for Code Exploration
+-- Spaghetti Comb - Neovim Plugin for Code Exploration
 -- Prevent loading if already loaded or if Neovim version is too old
-if vim.g.loaded_spaghetti_comb_v2 or vim.fn.has("nvim-0.8") ~= 1 then return end
-vim.g.loaded_spaghetti_comb_v2 = 1
+if vim.g.loaded_spaghetti_comb or vim.fn.has("nvim-0.8") ~= 1 then return end
+vim.g.loaded_spaghetti_comb = 1
 
 -- Task 11: User commands and keybindings
 
 -- UI Commands
 vim.api.nvim_create_user_command("SpaghettiCombBreadcrumbs", function()
-    require("spaghetti-comb-v2.ui.breadcrumbs").toggle()
+    require("spaghetti-comb.ui.breadcrumbs").toggle()
 end, { desc = "Toggle breadcrumb trail view" })
 
 vim.api.nvim_create_user_command("SpaghettiCombTree", function()
-    require("spaghetti-comb-v2.ui.floating_tree").toggle()
+    require("spaghetti-comb.ui.floating_tree").toggle()
 end, { desc = "Toggle navigation tree view with preview" })
 
 vim.api.nvim_create_user_command("SpaghettiCombBookmarks", function()
-    require("spaghetti-comb-v2.ui.picker").show_bookmark_mode()
+    require("spaghetti-comb.ui.picker").show_bookmark_mode()
 end, { desc = "Show bookmark picker" })
 
 vim.api.nvim_create_user_command("SpaghettiCombHistory", function()
-    require("spaghetti-comb-v2.ui.picker").show_navigation_mode()
+    require("spaghetti-comb.ui.picker").show_navigation_mode()
 end, { desc = "Show navigation history picker" })
 
 -- Bookmark Commands
 vim.api.nvim_create_user_command("SpaghettiCombBookmarkToggle", function()
-    local bookmarks = require("spaghetti-comb-v2.history.bookmarks")
+    local bookmarks = require("spaghetti-comb.history.bookmarks")
     local pos = vim.api.nvim_win_get_cursor(0)
     local location = {
         file_path = vim.api.nvim_buf_get_name(0),
@@ -35,7 +35,7 @@ vim.api.nvim_create_user_command("SpaghettiCombBookmarkToggle", function()
 end, { desc = "Toggle bookmark at current location" })
 
 vim.api.nvim_create_user_command("SpaghettiCombBookmarkClear", function(opts)
-    local bookmarks = require("spaghetti-comb-v2.history.bookmarks")
+    local bookmarks = require("spaghetti-comb.history.bookmarks")
     local global = opts.args == "global"
     bookmarks.clear_all_bookmarks(global)
     vim.notify("Bookmarks cleared", vim.log.levels.INFO)
@@ -47,7 +47,7 @@ end, {
 
 -- History Management Commands
 vim.api.nvim_create_user_command("SpaghettiCombHistoryClear", function(opts)
-    local history_manager = require("spaghetti-comb-v2.history.manager")
+    local history_manager = require("spaghetti-comb.history.manager")
     local project = opts.args ~= "all"
     -- Implementation would clear history
     vim.notify("History clearing not yet fully implemented", vim.log.levels.WARN)
@@ -59,7 +59,7 @@ end, {
 
 -- Debug Commands
 vim.api.nvim_create_user_command("SpaghettiCombDebug", function(opts)
-    local debug = require("spaghetti-comb-v2.utils.debug")
+    local debug = require("spaghetti-comb.utils.debug")
     if opts.args == "history" then
         debug.dump_history()
     elseif opts.args == "bookmarks" then
@@ -76,7 +76,7 @@ end, {
 })
 
 vim.api.nvim_create_user_command("SpaghettiCombLogLevel", function(opts)
-    local debug = require("spaghetti-comb-v2.utils.debug")
+    local debug = require("spaghetti-comb.utils.debug")
     local level = opts.args
     if level and level ~= "" then
         debug.set_log_level(level)
@@ -92,7 +92,7 @@ end, {
 
 -- Statusline Integration
 vim.api.nvim_create_user_command("SpaghettiCombStatus", function()
-    local statusline = require("spaghetti-comb-v2.ui.statusline")
+    local statusline = require("spaghetti-comb.ui.statusline")
     local status = statusline.get_branch_status()
     if status then
         vim.notify(string.format("Branch: %s, Depth: %d/%d, State: %s", status.branch_id, status.depth, status.total, status.state), vim.log.levels.INFO)
@@ -109,18 +109,18 @@ vim.api.nvim_create_user_command("SpaghettiCombSetup", function(opts)
         local ok, result = pcall(function() return vim.json.decode(opts.args) end)
         if ok then config = result end
     end
-    require("spaghetti-comb-v2").setup(config)
-    vim.notify("Spaghetti Comb v2 initialized", vim.log.levels.INFO)
+    require("spaghetti-comb").setup(config)
+    vim.notify("Spaghetti Comb initialized", vim.log.levels.INFO)
 end, {
     nargs = "?",
-    desc = "Setup Spaghetti Comb v2",
+    desc = "Setup Spaghetti Comb",
 })
 
 -- Task 12: Persistence commands
 
 -- Save current project history to disk
 vim.api.nvim_create_user_command("SpaghettiCombSaveHistory", function()
-    local history_manager = require("spaghetti-comb-v2.history.manager")
+    local history_manager = require("spaghetti-comb.history.manager")
     local success, msg = history_manager.save_current_project_history()
     local level = success and vim.log.levels.INFO or vim.log.levels.WARN
     vim.notify(msg or "Unknown error", level)
@@ -128,7 +128,7 @@ end, { desc = "Save navigation history for current project" })
 
 -- Save all project histories
 vim.api.nvim_create_user_command("SpaghettiCombSaveAllHistory", function()
-    local history_manager = require("spaghetti-comb-v2.history.manager")
+    local history_manager = require("spaghetti-comb.history.manager")
     local success, msg = history_manager.save_all_histories()
     local level = success and vim.log.levels.INFO or vim.log.levels.WARN
     vim.notify(msg or "Unknown error", level)
@@ -136,8 +136,8 @@ end, { desc = "Save navigation history for all projects" })
 
 -- Load project history from disk
 vim.api.nvim_create_user_command("SpaghettiCombLoadHistory", function()
-    local history_manager = require("spaghetti-comb-v2.history.manager")
-    local project_utils = require("spaghetti-comb-v2.utils.project")
+    local history_manager = require("spaghetti-comb.history.manager")
+    local project_utils = require("spaghetti-comb.utils.project")
 
     local current_file = vim.api.nvim_buf_get_name(0)
     local project_root = project_utils.detect_project_root(current_file)
@@ -154,7 +154,7 @@ end, { desc = "Load navigation history for current project" })
 
 -- Save current project bookmarks
 vim.api.nvim_create_user_command("SpaghettiCombSaveBookmarks", function()
-    local bookmarks = require("spaghetti-comb-v2.history.bookmarks")
+    local bookmarks = require("spaghetti-comb.history.bookmarks")
     local success, msg = bookmarks.save_current_project_bookmarks()
     local level = success and vim.log.levels.INFO or vim.log.levels.WARN
     vim.notify(msg or "Unknown error", level)
@@ -162,7 +162,7 @@ end, { desc = "Save bookmarks for current project" })
 
 -- Save all project bookmarks
 vim.api.nvim_create_user_command("SpaghettiCombSaveAllBookmarks", function()
-    local bookmarks = require("spaghetti-comb-v2.history.bookmarks")
+    local bookmarks = require("spaghetti-comb.history.bookmarks")
     local success, msg = bookmarks.save_all_bookmarks()
     local level = success and vim.log.levels.INFO or vim.log.levels.WARN
     vim.notify(msg or "Unknown error", level)
@@ -170,8 +170,8 @@ end, { desc = "Save bookmarks for all projects" })
 
 -- Load project bookmarks from disk
 vim.api.nvim_create_user_command("SpaghettiCombLoadBookmarks", function()
-    local bookmarks = require("spaghetti-comb-v2.history.bookmarks")
-    local project_utils = require("spaghetti-comb-v2.utils.project")
+    local bookmarks = require("spaghetti-comb.history.bookmarks")
+    local project_utils = require("spaghetti-comb.utils.project")
 
     local current_file = vim.api.nvim_buf_get_name(0)
     local project_root = project_utils.detect_project_root(current_file)
@@ -188,7 +188,7 @@ end, { desc = "Load bookmarks for current project" })
 
 -- Clean up old persistence files
 vim.api.nvim_create_user_command("SpaghettiCombCleanupPersistence", function(opts)
-    local storage = require("spaghetti-comb-v2.history.storage")
+    local storage = require("spaghetti-comb.history.storage")
     local max_age_days = opts.args and tonumber(opts.args) or 90
 
     local success, msg = storage.cleanup_persistence(max_age_days)
@@ -201,7 +201,7 @@ end, {
 
 -- Show storage statistics
 vim.api.nvim_create_user_command("SpaghettiCombStorageStats", function()
-    local storage = require("spaghetti-comb-v2.history.storage")
+    local storage = require("spaghetti-comb.history.storage")
     local stats = storage.get_storage_stats()
 
     local msg = string.format(
@@ -217,7 +217,7 @@ end, { desc = "Show persistence storage statistics" })
 
 -- List saved projects
 vim.api.nvim_create_user_command("SpaghettiCombListProjects", function()
-    local storage = require("spaghetti-comb-v2.history.storage")
+    local storage = require("spaghetti-comb.history.storage")
     local projects = storage.list_saved_projects()
 
     if #projects == 0 then
