+++ scripts/gendoc.lua
@@ -1,4 +1,4 @@
--- Documentation generation script for SpaghettiCombv2
+-- Documentation generation script for SpaghettiComb
 local MiniDoc = require("mini.doc")
 
 -- Setup mini.doc with default configuration
@@ -6,10 +6,10 @@ MiniDoc.setup()
 
 -- Define input files to process
 local input_files = {
-    "lua/spaghetti-comb-v2/init.lua",
+    "lua/spaghetti-comb/init.lua",
 }
 
-local output_file = "doc/spaghetti-comb-v2.txt"
+local output_file = "doc/spaghetti-comb.txt"
 
 -- Generate the documentation
 MiniDoc.generate(input_files, output_file)
